+++ library/src/main/java/com/patloew/rxlocation/RxLocationBaseOnSubscribe.java
@@ -9,6 +9,7 @@ import com.google.android.gms.common.api.PendingResult;
 import com.google.android.gms.common.api.Result;
 import com.google.android.gms.common.api.ResultCallback;
 import com.google.android.gms.common.api.Scope;
+import com.google.android.gms.location.ActivityRecognition;
 import com.google.android.gms.location.LocationServices;
 
 import java.util.concurrent.TimeUnit;
@@ -41,7 +42,7 @@ abstract class RxLocationBaseOnSubscribe<T> {
 
     protected RxLocationBaseOnSubscribe(@NonNull RxLocation rxLocation, Long timeout, TimeUnit timeUnit) {
         this.ctx = rxLocation.ctx;
-        this.services = new Api[]{LocationServices.API};
+        this.services = new Api[]{ LocationServices.API, ActivityRecognition.API };
         this.scopes = null;
 
         if (timeout != null && timeUnit != null) {