+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version 1.0.2
+
+* Update dependencies (RxJava 2.0.7, Play Services 10.2.0)
+* Fix bug which made activity recognition API unusable (#24)
+* Add `StatusExceptionResumeNextTransformer` to help with handlig the resolution of a `Result` yourself (#17)
+
 ## Version 1.0.1
 
 * Add consumer ProGuard rules. (#7, #8 – thanks Evisceration)